PIM-6352: exract cursor logic into an abstract
